Made LiveSelection a private class.
